edit amount api and implementation

+++ app/src/main/java/com/rino/self_services/model/pojo/payment/EditAmountRequest.kt
@@ -3,4 +3,4 @@ package com.rino.self_services.model.pojo.payment
 import com.google.gson.annotations.SerializedName
 
 data class EditAmountRequest (
-    @SerializedName("newAmount" ) var newAmount : Int?  = null)
+    @SerializedName("newAmount" ) var newAmount : Double?  = null)
